Read every generation of a stored code, and mean the type we declare
A column outlives the release that filled it, so reading an id now means
reading every id it may hold: today's packed code, the rank an ordinal
release stored, and the bytes themselves read as a code under any
padding or width this family has used -- with a superseded spelling
resolving through its aliases, so a store written ORDR still names
ORDER. Only bytes a generation would actually have written answer, so
from_int stays the strict reader a pushed code-set filter agrees with.
The currency CCCn bridge becomes one more case of that rule rather than
a branch of its own, the registry loader reads pairs and bare ids the
same way, and a value read back through a declaration heals instead of
degrading to UNKNOWN.

band moved onto the shared base but only some vocabularies band: one
that ranks each member by its own packed code declares no floors, so it
answers with itself rather than raising.

into_arrow_type promised a dictionary of the readable codes that nothing
could build -- Arrow indexes a dictionary by position, not by the value
stored. into_arrow_array now renders a stored column into exactly that
type, and the identity fallback refuses loudly instead of allocating
millions of entries for codes that were never positions.

The contracts go to version 2: their codes and three column widths
changed, and a version 1 Iceberg table needs those columns widened
before a version 2 producer can append to it.

+++ python/src/rekep/text/text_file.py
@@ -671,7 +671,7 @@ class TextFile(Dataset, io.BufferedIOBase):
             "eunix": pyarrow.nulls(count, pyarrow.int64()),
             "sunix": pyarrow.nulls(count, pyarrow.int64()),
             "version": _zeros(count, pyarrow.int64()),
-            "state": _zeros(count, pyarrow.int32()),
+            "state": _zeros(count, pyarrow.int64()),
             "code": pyarrow.repeat("", count),
             "codes": pyarrow.repeat(pyarrow.scalar({}, CODES_TYPE), count),
             "prev_unix": pyarrow.nulls(count, pyarrow.int64()),
